More refactoring: Restructured files and start breaking down components in to smaller parts. Start seperating styles from components... I'd like to do more but I think I've run out of time
