+++ src/components/AnimatedShapes/AnimatedShapes.js
@@ -1,7 +1,8 @@
 import React from 'react'
-import styled from 'styled-components'
 
-export default props => {
+import ImageContainer from './styles/ImageContainer'
+
+const AnimatedShapes = props => {
   // TODO: update width and height on window resize
   const clientWidth = React.useRef(document.body.clientWidth)
   const clientHeight = React.useRef(document.body.clientHeight)
@@ -13,8 +14,6 @@ export default props => {
   const maxDegreesY =  props.maxDegreesY || 25
 
   React.useEffect(() => {    
-    // TODO: This is not the best... I'd rather be using a ref to a 
-    // dom node than looking it up with a queryselector.
     const elementToListenOn = listenOnElementId
       ? document.querySelector(`#${listenOnElementId}`)
       : null
@@ -63,43 +62,4 @@ export default props => {
   )
 }
 
-// TODO: refine proportions (tranzlateZ, scale etc) to be true 
-// to original design 
-const ImageContainer = styled.div`
-  width: ${p => p.width || '585'}px;
-  height: ${p => p.height || '585'}px;;
-  position: relative;
-
-  img {
-    --rotateX: 0;
-    --rotateY: 0;
-    --translateZ: 0;
-    --scale: 1;
-    transform: 
-      perspective(600px) 
-      rotateY(var(--rotateX)) 
-      rotateX(var(--rotateY)) 
-      translateZ(var(--translateZ)) 
-      scale(var(--scale));
-    transition: transform 0.9s ease-out;
-    
-    position: absolute;
-    top: 0;
-    left: 0;
-  }
-  
-  img:nth-of-type(1) {
-    --translateZ: 150px; 
-    --scale: 0.9;
-  }
-  
-  img:nth-of-type(2) {
-    --translateZ: 40px; 
-    --scale: 0.95;
-  }
-  
-  img:nth-of-type(3) {
-    --translateZ: -80px;
-    --scale: 1;
-  }
-`
+export default AnimatedShapes
