+++ src/components/PrivateRoute.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Route, Redirect } from 'react-router-dom'
 
 import { useUserContext } from '../context/user/UserContext'
-import Loading from './Loading'
+import Loading from './common/Loading'
 
 const PrivateRoute = ({ children, ...rest }) => {
   const { state: user } = useUserContext()
