+++ src/components/Results/Pagination/styles/PagesCount.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+import SmallCaps from '../../../common/styles/SmallCaps'
+
+const PagesCount = styled(SmallCaps)`
+  margin: 0;
+`
+
+export default PagesCount
