+++ src/components/Results/ResultsList/styles/ResultsContainer.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const ResultsContainer = styled.div`
+  margin-top: 30px;
+  display: flex;
+  flex-wrap: wrap;
+`
+
+export default ResultsContainer
