+++ src/components/common/styles/SmallCaps.js
