+++ src/hooks/useGetResults.js
@@ -19,7 +19,13 @@ const useGetResults = () => {
 
   const { id: userId, username } = user
 
-  useEffect(() => {
+  useEffect(() => {    
+    /* 
+      Note: I thought about using useCallback for these and moving them outside the 
+      useEffect, but I think they are small enouggh functions that it's not entirely worth it 
+      should do some more reading on this to properly weight the performance cost/benefit  
+    */
+
     const getFavourites = async () => {
       try {
         const payload = await api.getFavourites(username, page)
